fsnotify Ops - Adding support for fil closure events Ticket #: EM-1861
Root Cause Analysis: fsnotify does not provide event notification when a file that was
opened for write is closed and actually modified.

Fix/Implementation: Add support for CLose_Write and Close_NoWrite events.  Also separate out
Modify from Write.  Modify is triggered when a file is actually modified and Write can be
triggered when a file is simply opened for write.

Reviewed By: <Name of Reviewer(s)>